Port URLEncoderTest
